refactor(dashboard): the additional address sits with the port it is not
It went to the overview because that is where the hostnames are, which reads as
an argument until you notice the run command is not there either. Both ports an
app has now sit together, and both lines worth copying off the page sit together.

It also stops a second card fetching what the first already had: the run command
reads the newest deployment, so the config card was holding this all along.
